Добавляет разбиение на модули
